refactor: moved MyCars from ContentView to its own View + added queryCar to DetailView
extracted StateObjects
